gurumitra-ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0
python-multipart==0.0.6
numpy>=1.26.4,<2.7.0
openai-whisper==20231117
yt-dlp>=2024.1.0
google-genai>=1.0.0
